proxy is now aware of webdav responses
